refactor(adapter-tests): invert bf-p inventory dependency — adapters register themselves
The inventory script hard-coded every adapter's factory/render/
not-available classes inside adapter-tests, the reverse of the package's
own architecture (run-adapter-conformance provides the constraint; each
adapter's test file registers itself). The driver moves to
src/hydration-props-inventory.ts knowing no adapter but the Hono
reference (the established exception — data-point-conformance,
generate-expected-html), exporting runHydrationPropsInventory +
HydrationPropsInventorySpec; each adapter package hosts a thin
scripts/hydration-props-inventory.ts wiring itself in, mirroring the
runAdapterConformanceTests registration shape. The adapter
devDependencies this PR had added to adapter-tests are reverted.
