+++ packages/adapter-tests/package.json
@@ -18,15 +18,10 @@
     "@barefootjs/shared": "workspace:*"
   },
   "devDependencies": {
-    "@barefootjs/blade": "workspace:*",
     "@barefootjs/client": "workspace:*",
-    "@barefootjs/erb": "workspace:*",
     "@barefootjs/go-template": "workspace:*",
     "@barefootjs/hono": "workspace:*",
-    "@barefootjs/jinja": "workspace:*",
     "@barefootjs/mojolicious": "workspace:*",
-    "@barefootjs/rust": "workspace:*",
-    "@barefootjs/twig": "workspace:*",
     "@barefootjs/xslate": "workspace:*",
     "@playwright/test": "^1.57.0",
     "happy-dom": "^20.11.1",
