+++ packages/adapter-tests/src/index.ts
@@ -26,6 +26,10 @@ export {
 export type { TemplatePrimitiveInput } from './cases/template-primitives'
 export { runAdapterConformanceTests } from './run-adapter-conformance'
 export type { RunAdapterConformanceOptions } from './run-adapter-conformance'
+export { compareHydrationProps } from './hydration-props-conformance'
+export type { BfPDivergenceKind, HydrationPropsComparisonResult } from './hydration-props-conformance'
+export { runHydrationPropsInventory } from './hydration-props-inventory'
+export type { HydrationPropsInventorySpec } from './hydration-props-inventory'
 export { assertDevReloadContract } from './dev-reload.contract'
 export type { DevReloadFacts } from './dev-reload.contract'
 export { assertScaffoldContract, ensureCreateCli } from './scaffold.contract'
